Sort settings/offline data/gpx tracks in reverse order.
There are two lists of GPX data in OSMAND.

1) Settings/offline data/GPX data
2) map/menu/define view/gpx track...

These lists are sorted in opposite orders.  This change flips the order of GPX data.

I prefer to see more recent items near the top.  It prevents scrolling to find a recent track.  And the most recent track is probably what the user wants to work on.